+++ examples/tutorial-01.rs
@@ -1,10 +1,9 @@
 #[macro_use]
 extern crate glium;
-extern crate glutin;
 
 fn main() {
     use glium::{DisplayBuild, Surface};
-    let display = glutin::WindowBuilder::new().build_glium().unwrap();
+    let display = glium::glutin::WindowBuilder::new().build_glium().unwrap();
 
     #[derive(Copy, Clone)]
     struct Vertex {