+++ tests/attrs.rs
@@ -1,6 +1,5 @@
 #[macro_use]
 extern crate glium;
-extern crate glutin;
 
 use glium::Surface;